+++ react/FilterNavigator.js
@@ -83,7 +83,7 @@ const FilterNavigator = ({
   fullWidthOnMobile = false,
   navigationTypeOnMobile = 'page',
   updateOnFilterSelectionOnMobile = false,
-  showClearByFilterOnMobile = false,
+  showClearByFilter = false,
 }) => {
   const { isMobile } = useDevice()
   const handles = useCssHandles(CSS_HANDLES)
@@ -207,7 +207,7 @@ const FilterNavigator = ({
               categoryFiltersMode={categoryFiltersMode}
               loading={loading}
               updateOnFilterSelectionOnMobile={updateOnFilterSelectionOnMobile}
-              showClearByFilterOnMobile={showClearByFilterOnMobile}
+              showClearByFilter={showClearByFilter}
             />
           </div>
         </div>