+++ react/components/AccordionFilterContainer.js
@@ -39,7 +39,7 @@ const AccordionFilterContainer = ({
   categoryFiltersMode,
   loading,
   onClearFilter,
-  showClearByFilterOnMobile,
+  showClearByFilter,
   updateOnFilterSelectionOnMobile,
 }) => {
   const intl = useIntl()
@@ -196,7 +196,7 @@ const AccordionFilterContainer = ({
                 truncatedFacetsFetched={truncatedFacetsFetched}
                 setTruncatedFacetsFetched={setTruncatedFacetsFetched}
                 onClearFilter={onClearFilter}
-                showClearByFilterOnMobile={showClearByFilterOnMobile}
+                showClearByFilter={showClearByFilter}
               />
             )
         }
@@ -244,7 +244,7 @@ AccordionFilterContainer.propTypes = {
   /** Clear filter function */
   onClearFilter: PropTypes.func,
   /** Whether a clear button that clear all options in a specific filter should appear beside the filter's name (true) or not (false). */
-  showClearByFilterOnMobile: PropTypes.bool,
+  showClearByFilter: PropTypes.bool,
   /** Wether the search will be updated on facet selection (`true`) or not (`false`) when the user is on mobile. */
   updateOnFilterSelectionOnMobile: PropTypes.bool,
 }