+++ react/components/AccordionFilterItem.js
@@ -35,7 +35,7 @@ const AccordionFilterItem = ({
   initiallyCollapsed,
   onClearFilter,
   facetKey,
-  showClearByFilterOnMobile,
+  showClearByFilter,
 }) => {
   const intl = useIntl()
   const handles = useCssHandles(CSS_HANDLES)
@@ -107,7 +107,7 @@ const AccordionFilterItem = ({
                   <Tag>{quantitySelected}</Tag>
                 </div>
               )}
-              {quantitySelected > 0 && showClearByFilterOnMobile && (
+              {quantitySelected > 0 && showClearByFilter && (
                 <span
                   className={classNames(
                     handles.accordionFilterItemTag,
@@ -182,7 +182,7 @@ AccordionFilterItem.propTypes = {
   /** Facet's key */
   facetKey: PropTypes.string,
   /** Whether a clear button that clear all options in a specific filter should appear beside the filter's name (true) or not (false). */
-  showClearByFilterOnMobile: PropTypes.bool,
+  showClearByFilter: PropTypes.bool,
 }
 
 export default AccordionFilterItem